+++ tests/common/experimental/test_statistic_collector.py
